rename pages
maybe clearer like this
