+++ src/pages/AllCollectionsPage/AllCollectionsPage.vue
