+++ src/pages/AllCollectionsPage/CollectionItem.vue
