+++ src/pages/BrowseCollectionPage/BrowseCollectionPage.vue
